update tpm context serialization routines
Partial fix for #229. Previously, the save and
load context from file routines were just casting
the TPMS_CONTEXT structure as a byte array and
writing it to disk. This has the negative side
effect of writing any padding, and not handling
the endianess of any internal fields. Thus,
the file output was architecture and compiler
specific.

Change this to use the new file io serialization routines
to write each field out to disk.

Update the interface to be consistent with the coding
standard as well.

Signed-off-by: William Roberts <william.c.roberts@intel.com>